Autotagging/requirements.txt
activation==0.1.0
appdirs==1.4.3
appnope==0.1.0
apturl==0.5.2
argh==0.26.2
asn1crypto==0.24.0
atomicwrites==1.2.1
attrs==18.2.0
backcall==0.1.0
backports-abc==0.5
base58==1.0.3
beautifulsoup4==4.6.3
bleach==2.0.0
blis==0.2.4
boto==2.49.0
boto3==1.9.92
botocore==1.12.92
Brlapi==0.6.6
bz2file==0.98
certifi==2019.3.9
chardet==3.0.4
citelementtree==1.2.7
Click==7.0
cloudant==2.11.0
cloudpickle==0.7.0
command-not-found==0.3
corenlp-protobuf==3.8.0
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
cymem==2.0.2
Cython==0.29
cytoolz==0.8.2
dask==1.1.0
decorator==4.0.11
defer==1.0.6
defusedxml==0.5.0
dill==0.2.8.2
distro-info==0.18
Django==2.1.7
docutils==0.14
elasticsearch==6.3.1
en-core-web-sm==2.0.0
entrypoints==0.2.2
enum-compat==0.0.2
Flask==0.10.1
Flask-Cors==3.0.7
flexmock==0.10.4
ftfy==4.4.3
future==0.17.1
future-fstrings==0.4.5
futures==3.1.1
gensim==3.7.1
graphviz==0.10.1
h5py==2.8.0
html5lib==0.999999999
httplib2==0.9.2
idna==2.8
image==1.5.27
imutils==0.5.2
ipykernel==4.5.2
ipython==5.3.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
itsdangerous==1.1.0
jedi==0.13.1
Jinja2==2.9.5
jmespath==0.9.3
jsonpickle==1.1
jsonrpclib==0.1.7
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.0.0
jupyter-console==5.1.0
jupyter-core==4.3.0
Keras==2.1.2
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
keyring==10.6.0
keyrings.alt==3.0
Kivy-Garden==0.1.4
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lightdm-gtk-greeter-settings==1.2.2
louis==3.5.0
lxml==4.2.5
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.0.1
MarkupSafe==1.0
matplotlib==2.0.0
mistune==0.7.4
mock==2.0.0
more-itertools==4.3.0
msgpack==0.5.6
msgpack-numpy==0.4.3.2
murmurhash==1.0.2
nbconvert==5.1.1
nbformat==4.3.0
networks==0.3.7
networkx==2.2
nltk==3.3
notebook==4.4.1
numpy==1.16.2
oauth==1.0.1
olefile==0.45.1
opencv-contrib-python==3.4.4.19
opencv-python==3.4.4.19
packaging==16.8
pandas==0.23.4
pandocfilters==1.4.1
parso==0.3.1
pathlib==1.0.1
pathtools==0.1.2
patsy==0.5.1
pbr==5.1.2
pexpect==4.2.1
pickleshare==0.7.4
Pillow==5.1.0
plac==0.9.6
pluggy==0.8.0
preshed==2.0.1
prometheus-client==0.4.2
prompt-toolkit==1.0.14
protobuf==3.2.0
psutil==5.4.7
ptyprocess==0.5.1
py==1.7.0
pycairo==1.16.2
pycorenlp==0.3.0
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.2.0
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.2.0
pyRFC3339==1.0
PyStanfordDependencies==0.3.1
pytesseract==0.2.6
pytest==3.9.1
python-apt==1.6.3
python-dateutil==2.6.0
python-debian==0.1.32
pytz==2017.2
PyWavelets==1.0.1
pyxattr==0.6.0
pyxdg==0.25
PyYAML==3.12
pyzmq==16.0.2
qtconsole==4.3.0
regex==2017.11.9
reportlab==3.4.0
requests==2.21.0
requests-unixsocket==0.1.5
s3transfer==0.2.0
scikit-image==0.14.2
scikit-learn==0.20.3
scipy==1.2.1
seaborn==0.9.0
SecretStorage==2.3.1
semver==2.8.1
Send2Trash==1.5.0
sequential==1.0.0
simplegeneric==0.8.1
simplejson==3.13.2
singledispatch==3.4.0.3
six==1.10.0
smart-open==1.8.0
spacy==2.1.3
sputnik==0.9.3
srsly==0.0.5
ssh-import-id==5.7
stanford-corenlp==3.8.0
stanfordcorenlp==3.9.1.1
stanfordnlp==0.1.2
statsmodels==0.9.0
streamlit==0.25.0
system-service==0.3
systemd-python==234
tensorflow==1.0.0
tensorflow-tensorboard==0.1.8
termcolor==1.1.0
terminado==0.6
tesseract==0.1.3
testpath==0.3
tflearn==0.3.2
thinc==7.0.4
tokenize-rt==2.1.0
toml==0.10.0
toolz==0.9.0
torch==0.4.1
torchvision==0.2.1
tornado==4.4.3
tqdm==4.31.1
traitlets==4.3.2
tzlocal==1.5.1
ubuntu-drivers-common==0.0.0
ufw==0.35
ujson==1.35
unattended-upgrades==0.1
urllib3==1.24.1
usb-creator==0.3.3
utils==0.9.0
versions==0.10.0
wadllib==1.3.2
wasabi==0.2.1
watchdog==0.9.0
wcwidth==0.1.7
webencodings==0.5
Werkzeug==0.14.1
widgetsnbextension==2.0.0
workbook==1.1
wrapt==1.10.11
xkit==0.0.0
xlrd==1.1.0
xlutils==2.0.0
xlwt==1.3.0
yellowbrick==0.9.1
youtube-dl==2018.3.14
zope.interface==4.3.2
